Add pandas dependency and mount logs
A specific version of pandas is required to run the merge_queries.py script. Pandas is a binary dependency and so cannot be included in Airflow packaged DAGs.
Mount the airflow logs server on the host otherwise the logs get lost any time the container is recreated (e.g if environment variables change).